// app/model/Item.ts
export class Item {/*
    undefined: string;
    __EMPTY: string;
    __EMPTY_1: string;
    __EMPTY_2: string;
    __EMPTY_3: string;
    __EMPTY_4: string;*/
     
        concept: string;
        date: string;
        paid: number;
        option: boolean;
        type: string;
    
}